Code review and changes in auth epic handling. Need to approve. Project structure changes.
