+++ frontend/react-client/src/app/config.ts
@@ -1,1 +1,3 @@
 export const API_URL = "https://localhost:7213";
+
+export const AUTH_KEY_NAME = '__AUTH';
