+++ frontend/react-client/src/app/gates/RequirePermission.tsx
@@ -1,8 +1,8 @@
-import useIsHavePermission from "../../hooks/useIsHavePermission.ts";
+import useIsHavePermission from "../hooks/useIsHavePermission.ts";
 import {Outlet} from "react-router-dom";
-import UnauthorizedPage from "../../pages/UnauthorizedPage.tsx";
+import UnauthorizedPage from "../ui/pages/UnauthorizedPage.tsx";
 
-export default function RequirePermission({permission}){
+export default function RequirePermission({permission}: {permission: string}){
     const isHavePermission = useIsHavePermission(permission);
     return isHavePermission? <Outlet/> : <UnauthorizedPage/>
 }
