+++ frontend/react-client/src/app/hooks/useIsAuthenticated.ts
