+++ frontend/react-client/src/app/hooks/useIsHavePermission.ts
