+++ frontend/react-client/src/app/misc/RequestCreator.ts
@@ -1,0 +1,16 @@
+import { API_URL } from "../config";
+
+export const createRequest = (query: string) => {
+  const request = {
+    url: `${API_URL}/graphql`,
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: {
+        query: query,
+      },
+}
+    return request; 
+
+  };
