+++ frontend/react-client/src/app/store.ts
@@ -1,0 +1,29 @@
+import {combineEpics, createEpicMiddleware, Epic} from "redux-observable";
+import {Action, configureStore, Tuple} from "@reduxjs/toolkit";
+
+import authReducer from "./features/authentification/authSlice.ts";
+import { authUserEpic } from "./features/authentification/epics.ts";
+import { useDispatch, useSelector } from "react-redux";
+
+
+const rootEpic: Epic<Action, Action, void, any> = combineEpics<Action, Action, void, any>(
+    authUserEpic
+  );
+
+const epicMiddleware = createEpicMiddleware<Action, Action, void, any>();
+
+
+export const store = configureStore({
+    reducer: {
+        auth: authReducer
+    },
+    middleware: () => new Tuple(epicMiddleware)
+})
+
+epicMiddleware.run(rootEpic);
+
+export const useAppDispatch = useDispatch.withTypes<AppDispatch>()
+export const useAppSelector = useSelector.withTypes<RootState>()
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
