+++ frontend/react-client/src/app/types/User.ts
@@ -1,4 +1,4 @@
-type User={
+interface User {
     id: number,
     fullName: string,
     email: string,
