+++ frontend/react-client/src/app/ui/pages/AccountVerificationPage.tsx
@@ -5,30 +5,35 @@ import Button from '@mui/material/Button';
 import TextField from '@mui/material/TextField';
 import Box from '@mui/material/Box';
 import Grid from '@mui/material/Grid';
-import AddIcon from '@mui/icons-material/Add';
+import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
 import Typography from '@mui/material/Typography';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { Link as RouterLink } from 'react-router-dom';
 import { Link as MuiLink } from '@mui/material';
 
-
 const defaultTheme = createTheme();
-const RegisterPage: React.FC = () => {
+const AccountVerificationPage: React.FC = () => {
     const [error, setError] = useState<string>('');
 
-    const handleSubmit = (event) => {
+    const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
         const data = new FormData(event.currentTarget);
 
-        if (data.get('name') != '' || data.get('email') != '') {
+        if (data.get('code') !== '' && data.get('password') !== '' && data.get('repeatPassword') !== '') {
             console.log({
-                email: data.get('name'),
-                password: data.get('email'),
+                email: data.get('code'),
+                password: data.get('password'),
+                repeatPassword: data.get('repeatPassword')
             });
         }
 
-        if (data.get('name') === '' || data.get('email') === '') {
+        if (data.get('password') === '' || data.get('repeatPassword') === '' || data.get('code') === '') {
             setError('Please fill in all fields');
+            return;
+        }
+
+        if (data.get('password') !== data.get('repeatPassword')) {
+            setError('Passwords do not match');
             return;
         }
     };
@@ -45,33 +50,43 @@ const RegisterPage: React.FC = () => {
                 }}
             >
                 <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
-                    <AddIcon />
+                    <LockOutlinedIcon />
                 </Avatar>
                 <Typography component="h1" variant="h5">
-                    Create User
+                    Verification
                 </Typography>
                 <Box component="form" noValidate onSubmit={handleSubmit} sx={{ mt: 1 }}>
                     <TextField
                         margin="normal"
                         required
                         fullWidth
-                        id="name"
-                        label="Name"
-                        name="name"
-                        autoComplete="name"
+                        id="code"
+                        label="Verification Code"
+                        name="code"
+                        autoComplete="code"
                         autoFocus
                     />
                     <TextField
                         margin="normal"
                         required
                         fullWidth
-                        name="email"
-                        label="Email Address"
-                        type="email"
-                        id="email"
-                        autoComplete="email"
+                        name="password"
+                        label="Password"
+                        type="password"
+                        id="password"
+                        autoComplete="current-password"
                     />
-                    <Typography color="error"x>
+                    <TextField
+                        margin="normal"
+                        required
+                        fullWidth
+                        name="repeatPassword"
+                        label="Repeat Password"
+                        type="password"
+                        id="repeatPassword"
+                        autoComplete="current-password"
+                    />
+                    <Typography color="error">
                         {error}
                     </Typography>
                     <Button
@@ -80,12 +95,17 @@ const RegisterPage: React.FC = () => {
                         variant="contained"
                         sx={{ mt: 1, mb: 2 }}
                     >
-                        Create User
+                        Verification
                     </Button>
+                    <Grid item>
+                        <MuiLink component={RouterLink} to="/login" variant="body2">
+                            {"Do you have an account? Sign In"}
+                        </MuiLink>
+                    </Grid>
                 </Box>
             </Box>
         </ThemeProvider>
     );
 };
 
-export default RegisterPage;
+export default AccountVerificationPage;
