+++ frontend/react-client/src/app/ui/pages/ProtectedPage.tsx
