+++ frontend/react-client/src/app/ui/pages/RegisterPage.tsx
@@ -3,43 +3,31 @@ import {useState} from "react";
 import Avatar from '@mui/material/Avatar';
 import Button from '@mui/material/Button';
 import TextField from '@mui/material/TextField';
-import Link from '@mui/material/Link';
 import Box from '@mui/material/Box';
-import Grid from '@mui/material/Grid';
-import PersonIcon from '@mui/icons-material/Person';
+import AddIcon from '@mui/icons-material/Add';
 import Typography from '@mui/material/Typography';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
-import {Link as RouterLink, Navigate} from 'react-router-dom';
-import { Link as MuiLink } from '@mui/material';
-import {useDispatch} from "react-redux";
-import {loginUser} from "../state/auth/epics.ts";
-import useIsAuthenticated from "../hooks/useIsAuthenticated.ts";
-
 
 const defaultTheme = createTheme();
-const LoginPage: React.FC = () => {
+const RegisterPage: React.FC = () => {
     const [error, setError] = useState<string>('');
-    const isAuthenticated = useIsAuthenticated();
-    const dispatch = useDispatch();
-    const handleSubmit = (event) => {
+
+    const handleSubmit = (event: React.FormEvent<HTMLFormElement>): void => {
         event.preventDefault();
         const data = new FormData(event.currentTarget);
 
-        const email = data.get('email');
-        const password = data.get('password');
+        if (data.get('name') !== '' || data.get('email') !== '') {
+            console.log({
+                email: data.get('name'),
+                password: data.get('email'),
+            });
+        }
 
-        if (email === '' || password === '') {
+        if (data.get('name') === '' || data.get('email') === '') {
             setError('Please fill in all fields');
             return;
         }
-
-        console.log({email, password});
-
-        dispatch(loginUser({email, password}))
     };
-
-    if(isAuthenticated)
-        return <Navigate to={"/"}/>
 
     return (
         <ThemeProvider theme={defaultTheme}>
@@ -53,33 +41,33 @@ const LoginPage: React.FC = () => {
                 }}
             >
                 <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
-                    <PersonIcon />
+                    <AddIcon />
                 </Avatar>
                 <Typography component="h1" variant="h5">
-                    Sign in
+                    Create User
                 </Typography>
                 <Box component="form" noValidate onSubmit={handleSubmit} sx={{ mt: 1 }}>
                     <TextField
                         margin="normal"
                         required
                         fullWidth
-                        id="email"
-                        label="Email Address"
-                        name="email"
-                        autoComplete="email"
+                        id="name"
+                        label="Name"
+                        name="name"
+                        autoComplete="name"
                         autoFocus
                     />
                     <TextField
                         margin="normal"
                         required
                         fullWidth
-                        name="password"
-                        label="Password"
-                        type="password"
-                        id="password"
-                        autoComplete="current-password"
+                        name="email"
+                        label="Email Address"
+                        type="email"
+                        id="email"
+                        autoComplete="email"
                     />
-                    <Typography color="error">
+                    <Typography component="span" color="error">
                         {error}
                     </Typography>
                     <Button
@@ -88,24 +76,12 @@ const LoginPage: React.FC = () => {
                         variant="contained"
                         sx={{ mt: 1, mb: 2 }}
                     >
-                        Sign In
+                        Create User
                     </Button>
-                    <Grid container>
-                        <Grid item xs>
-                            <Link href="#" variant="body2">
-                                Forgot password?
-                            </Link>
-                        </Grid>
-                        <Grid item>
-                            <MuiLink component={RouterLink} to="/verification" variant="body2">
-                                {"Account verification"}
-                            </MuiLink>
-                        </Grid>
-                    </Grid>
                 </Box>
             </Box>
         </ThemeProvider>
     );
 };
 
-export default LoginPage;
+export default RegisterPage;
