--- frontend/react-client/src/state/rootEpic.ts
@@ -1,8 +1,0 @@
-import {combineEpics} from "redux-observable";
-import {fetchHelloEpic} from "./hello/epics.ts";
-import {loginUserEpic} from "./auth/epics.ts";
-
-export const rootEpic = combineEpics(
-    loginUserEpic,
-    fetchHelloEpic
-)
